requirements.txt
openai
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
